// src/app/api/models/dashboard-metrics-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AppointmentMetricsDto } from '../models/appointment-metrics-dto';
import { BusinessTypeMetricsDto } from '../models/business-type-metrics-dto';
import { ClientMetricsDto } from '../models/client-metrics-dto';
import { ConsultationMetricsDto } from '../models/consultation-metrics-dto';
import { RevenueMetricsDto } from '../models/revenue-metrics-dto';
export interface DashboardMetricsDto {
  appointments: AppointmentMetricsDto;
  businessTypes: Array<BusinessTypeMetricsDto>;
  clients: ClientMetricsDto;
  consultations: ConsultationMetricsDto;
  generatedAt: string;
  revenue: RevenueMetricsDto;
}